Fixed the deadlock in ManagedPoolTest.
Though I could never get the deadlock to occur on MacOSX, it *always*
happened on any Linux box I tried (both Fedora and Ubuntu).  My first
attempts at inserting S.o.println debug statements would remove the
deadlock, making the race condition harder to track, but I ultimately
found it by putting them in the *right* place.

Essentially, the deadlock occurred in PoolManager.instanceRequested
whenever the inFlight member was true.  When this happened,
SimplePool.fillInstance would never get called and
SimplePool.borrowInstance would wait forever, hence deadlock.

Looking at the code, it seemed to me that the inFlight member is
unnecessary in light of 1) the instances semaphore ensuring an upper
bound and 2) the single-threaded executor ensuring that the waiting
borrowers will eventually get their instance.

So I removed inFlight, and that fixed the unit tests.

If it turns out I'm a total dumbshit and missed something
obvious (concurrency is hard!) that requires inFlight, we need to 1)
add a test for that obvious thing, and 2) make sure we test it on
Linux.  ;)