Ok, sort of fixing TORQUE-30.
Renamed torquebox-rails to rake-support and made the assembly depend
on it, making our torquebox/tasks available once again.

This still doesn't account for the rails template, however, since I'm
not sure into which sub-project bin.xml should be ressurected.  So you
can refer to the template from your local maven repo, ala:

  rails -m ~/.m2/repository/org/torquebox/rake-support/1.0.0.Beta21/rake-support-1.0.0.Beta21.template.rb your-test-app

Still not happy with some of the names here, though, and the fact that
we have source files not beneath src/ that don't appear to be used,
e.g. init.rb and lib/torquebox-rails.rb.