This is probably stupid.
I had to manually install the activerecord-jdbcsqlite3-adapter to get
a virgin rails app to run.  I figured since that guy would pull in the
activerecord-jdbc-adapter as one of its own deps, I figure why not
kill two birds and just depend on the sqlite3 adapter.  Again,
probably stupid, but hey it eliminates the user's surprise that he
needs to install it himself when the server:run task fails, right?