This fixes TORQUE-95, but I'm not sure it's correct.
I feel like the result of File.expand_path when the 2nd parameter is a
vfs url should be another vfs url, but the sqlite jdbc driver won't
like that at all.  So we return a "real" file path, and hope nobody
ever packages their database in a file within a war.  :)