Bind tested naming service to an available port.
Prevents port conflicts with JBoss if it happens to be running at the
time you run your tests.  The available_port method isn't ideal -- I'd
prefer to just set the naming service (and rmi) ports to 0, but it
wasn't obvious to me how to query them for the actual ports returned
by the OS.  Regardless, the tests pass with jboss running now. :)