Essentially simplified load path maintenance.
Additional load path elements may be configured using the new
LoadPathDeployer, which replaces both the
BasicSubsystemConfigurationDeployer and the
SubsystemLoadPathDescriber.

We removed the SubsystemConfiguration framework for now, as all it
really did was munge the load path, in a somewhat complex way.  The
complexity was in anticipation of more configuration requirements, so
we may bring it back once they're known.

Removing the SubsystemConfiguration stuff left no justification for
the existence of base-int, so it's gone now.  One less pom.xml!