Quick (partial) fix for TORQUE-107.
This could stand some better injection, similar to what we do for
message processors.

It only partially fixes 107 because only changes to the job class are
reloaded.  References to model classes are not.  I confirmed that
tasks have the same problem, so I'll create a new feature request that
addresses both of them.