Simplifying the rubyConfig accessors for MessageProcessors.
I'm not entirely sure why we were marshalling this to a byte[] before,
since JRuby seems to handle converting Hashs to Maps transparently, so
I just changed the accessors to be of type Map, and everything still
seems to work fine.  It allowed me to remove a lot of code, and makes
converting the metadata_builder.rb to Java more straightforward.