Moving the encode/decode logic from clients to messaging-core.
This is so that the MessageProcessor can take advantage of the same
logic as the client.  This means that the thing passed to on_message
may not be a String.  It'll be whatever type the Ruby producer passed
as its message.  This'll be the place we expand to handle other JMS
message subtypes, too.