Here's a pretty darn clever integration test... that doesn't work.
The test will issue a web request that will publish a message to
/queues/test.  The test will then block on a receive of a message from
/queues/results.  A message processor configured in the app will
receive the message on /queues/test and publish it to /queues/results,
theoretically satisfying the test.

Unfortunately the test bombs when it tries to load the
org.torquebox.torquebox-messaging-client gem.  This problem is similar
to what we saw with our rspec experiments when the vfs gem could not
be loaded.  I've tried everything I know to do, but I'm obviously
doing it wrong, so I kinda need help here.  :(