Fixes TORQUE-148
We no longer create the "web" PoolMetaData in two different deployers,
AppRackYamlParsingDeployer and RailsRubyRuntimeFactoryDescriber.
Instead, we create/attach it only in RackApplicationPoolDeployer,
which makes sense since it'll be the chief consumer of the pool.
Further, we only create the pool if PoolingYamlParsingDeployer hasn't
already done so, allowing users to override the default web pool using
the pooling.yml internal deployment descriptor.

We also added a convenience method on AttachmentUtils to return a
multiply-attached object by name, using the same convention used when
it was attached.