Moving RubyRuntimeMetaData creation out of AppRailsYamlParsingDeployer.
It's already being created in the downstream
RailsRubyRuntimeFactoryDescriber, except for the environment variables
which, since we've added an attribute for them to
RackApplicationMetaData, it can now do.

It's a little weird that AppRailsYamlParsingDeployer produces both a
RailsApplicationMetaData *and* a RackApplicationMetaData, and hence
RailsRubyRuntimeFactoryDescriber now consumes both.  More refactoring
will come because we need rack apps to have internal descriptors for
environment variables and load paths, eventually specified in
torquebox.yml.