Moving MessagingRuntimePoolDeployer from PRE_REAL to DESCRIBE where it belongs.
In so doing, apps with tasks (but no messaging.yml or pooling.yml)
ceased to work, because the inputs and outputs weren't declared
correctly for MessagingRuntimePoolDeployer and TasksDeployer, which
were now in the same stage.  So obviously I needed to declare an input
on MessageProcessorMetaData on MessagingRuntimePoolDeployer.  And then
I hit on the bright idea of using setInput instead of addInput so as
to require the attached MessageProcessorMetaData.  This would allow me
to eliminate the getAllMetaData short circuit in deploy, which is
kinda ick.  Doing this broke a unit test, however.  WTF?  It seems the
setInput only works for metadata attached with an ExpectedType
parameter, which our AttachmentUtils doesn't do, and I probably
shouldn't assume that any upstream deployers would.  Still, kinda
makes the setInput a little useless and errorprone to use, IMHO.

Though not strictly necessary, I made similar changes to
JobsRuntimePoolDeployer, just in case a same-stage deployer ever
attaches a ScheduledJobMetaData.