Implement File.chmod for files inside archives(TORQUE-168)
If the files inside an archive are mounted in a real filesystem,
File.chmod will update the real files. This is used by some gems
that write temporary files to the Rails tmp/ directory.

Dir.mkdir was changed to use a new method, VFS.writable_path_or_error,
that File.chmod also uses to abstract away the behavior of either
returning a writable path to a real file or throwing an appropriate
error.