Caching works in archives now, woo-hoo!
I had to make one minor tweak due to the fact that rack meta data is
"write once", so it's not possible to reset rackRoot once it's been
set.  Otherwise, an internal descriptor could override what an
external descriptor sets.  So I added a new method,
explode(), (because nouns are the devil's handiwork) about which no
descriptor is aware.