There is both success and failure here.
My intent was to move some of the basic sinatra logic to rspec, mostly
to demonstrate some of the cooler methods in the DSL, but then I
noticed that the 500 posts test was dog slow in rspec, mostly because
capybara does its remote testing in an actual browser.  This will be
yet another thing to piss CI off.  There doesn't seem to be a way to
specify :htmlunit as the browser for the selenium driver to use,
though if you could you'd do it in that new code in spec_helper.rb.
Another option would be to use selenium's :remote browser option,
after starting up say the HtmlUnitDriver driver in another process.
Or we could use the celerity driver, but that doesn't seem to work
well with 0.4.0.  So we have a firefox window pop up during our rspec
tests for now.