Long-running services seem to be working! [TORQUE-192]
I changed the original design a tad.  Instead of all the
directory/file naming conventions, I opted for a simple "services.yml"
file, mainly to support jobs needing to be initialized with
parameters.  They're probably gonna get 'em from a yaml file anyway,
so it may as well be ours, right?

Anyhoo, top-level keys are the names of the service classes.  The
actual class files should reside somewhere on the load-path (either
lib/ or app/models/ for Rails or off the root of a Rack app).  The
value of each key in the YAML file should be an optional map (or list)
of parameters that we'll pass on to the service's initialize method.
The service should also implement :start and :stop.

More tests and docs forthcoming.