Fix for TORQUE-195.
We're not firing up a thread as proposed in the issue, because a bit
of research revealed that JMS sessions don't cotton to that.  Instead,
we're introducing a 'concurrency' attribute for the MessageProcessor
to limit the number of concurrent messages sent to a specific type of
processor.  The default is 1, i.e. single-threaded.