This fixes TORQUE-211, but it gives me the willies.
I don't like returning "/" in response to dirname("vfs:").  I'm also
not crazy about the fact that the resulting path of a Tempfile
constructed with a vfs path does NOT have a vfs path.  But since we
explode everything anyway, I don't think it matters.

I'm particularly uneasy about the fact that i don't understand why
this is broken.  I want to say it's something about the combination of
JRuby's native implementation of tempfile.rb coupled with the fact
that Tempfile extends DelegatorClass(File).  Something about that
wickedness leads to mucho unpredictability regarding monkey-patching
the new method.