Expose methods for synchronous messaging in Queues (TORQUE-235)
A client wanting to send a synchronous message to a consumer and
receive a synchronous response can call queue.publish_and_receive,
which accepts the same arguments as queue.publish. The consumer should
call queue.receive_and_publish, which accepts the same arguments as
queue.receive but also takes a block. The block is passed the message
from the client and its return value is the message that will get sent
back to the client as a response.

Below is n example that sends a message to a consumer and waits for
the consumer to respond with the uppercased text of that message.

Client:
queue = TorqueBox::Messaging::Queue.new("/queues/upcase")
queue.publish_and_receive("ping")

Consumer:
queue = TorqueBox::Messaging::Queue.new("/queues/upcase")
queue.receive_and_publish { |message| message.upcase }

If publish_and_receive is used within a web request, be sure to pass a
timeout option so the request won't hang indefinitely if the consumer
is taking too long to respond.

queue.publish_and_receive("ping", :timeout => 1000) # 1 second timeout

If publish_and_receive times out the return value will be nil.