Allow tasks on any object [TORQUE-213]
This change creates a queue (/queues/torquebox/<app name>/embedded-tasks) and
registers a processor for the queue that extracts a receiver from the message
and calls the specified method on the receiver.

On the ruby side, there is now a module (TorqueBox::Messaging::EmbeddedTasks)
that when extended gives the class a :handle_async method that marks a method
to be processed asynchronously using the embedded-tasks queue.

Usage:

class SomeClass
  extend TorqueBox::Messaging::EmbeddedTasks

  def a_method
  end
  handle_async :a_method
end

SomeClass.new.a_method # :a_method will actually be called asynchronously
by the embedded-tasks processor.