TORQUE-237: Rework queues.yml and topics.yml to use the torquebox.yml-based stuffs.
Each deployer will check first for an appropriate section supplied through torquebox.yml.
If a section is not present, then the appropriate specific *.yml is searched for.  If found,
the deployer gets the chance to cast the resulting Object to whatever it expects from its
section, and then create and attach metadata to the unit as appropriate.

Add tests to confirm behaviour when either/both sources of configuration information
is present, for topics.yml and queues.yml.

I hope Toby enjoys this comment.