Initial support for turning a Service into a HASingleton (TORQUE-194)
Setting "singleton: true" to a service entry in services.yml will now
make that service a HASingleton. This means that the service will be
instantiated on every node of the cluster but only started on the
master node. If the master node goes down, a new master node will be
elected and the service started on that node.

A HASingleton doesn't guarantee no downtime of the service - it just
ensures the service is only running on one node and starts the service
on a new node if the master node goes down. However, there will be a
small duration of time between the old master going down and a new
master being elected and the service being started.

The singleton key will not be passed in the options hash to the
initialize method of a service. So, treat singleton as a reserved key
and don't attempt to use that key when initializing the service.

Example of singleton usage in services.yml:

IrcBotService:
  singleton: true
  nick: torquebox_bot
  server: irc.freenode.net
  port: 6667
  channel: '#torquebox'