Expose hornetQ knobs for priority, ttl, and persistence [TORQUE-214]
These options are exposed at the Queue level, from Backgroundable, and
for regular async tasks. The options are passed as a hash, like so:

{
  :priority => :high, # 7
  :ttl => 5,          # seconds
  :persistent => true
}

The priority can be anything in the range 0..9, or one of the convenience
symbols:

{
  :low => 1,
  :normal => 4,
  :high => 7,
  :critical => 9
}

If any option is not provided, it defaults to the hornetQ default.

Usage:

For a Queue:

Queue.new(queue_name).publish(message, :priority => :critical)

For an async task:

ATask.async(:method, payload, :ttl => 0.5)

For #always_background:

AModel.always_background(:method, :persistent => false)

For #background:

an_object.background(:priority => 4).method

This patch retasks the options param to Queue.publish - it was used
as connection options for Client.connect, but is no longer. The options
passed to the Queue constructor are solely used instead.