Store session data in a SessionData class (TORQUE-264, TORQUE-265)
This fixes two issues around ServletStore.

The first is that profiling showed every http thread was competing for
the same lock on org.jruby.util.ClassCache$OneShotClassLoader and thus
reducing performance.

The second is that every request was causing a new class to load in
the JVM, eating through PermGen space. The classes were eventually
unloaded and the PermGen space reclaimed but it's a bit unsettling.