Queues and Topics will wait until deployed instead of erroring (TORQUE-266)
It's possible to call publish or receive on a queue or topic before
it's fully deployed if you do it early in the application (a service's
start method, on Rails app boot). Previously that would throw an error
but now the publish or receive call will block until the destination
is available.

The default maximum wait time is 30 seconds but can be overriden by
the :startup_timeout option:

queue.publish "my_message", :startup_timeout => 5000 # 5 seconds

If the startup timeout is exceeded and the destination still isn't
available the original javax.naming.NameNotFoundException will be
raised.