Exposed concurrency for app/tasks/ and Backgroundable [TORQUE-244, TORQUE-268]
You can now set the message processor concurrency for any task class queue and for
the Backgroundable queue in torquebox.yml:

messaging:
  /queues/not-part-of-this-example:
    SomeProcessor
  tasks:
    Backgroundable:
      concurrency: 5
    MySuperTask:
      concurrency: 10

Setting the Backgroundable concurrency to 0 prevents the Backgroundable queue from
being created, and any attempt to call a backgrounded method in the app will raise
a helpful exception.

Unfortunately, this change also allows you to set the concurrency to 0 for the
task class queues, or any message processor, which will prevent it from deploying.
We log a warning in that case.

I'm not happy at all with this implementation, since it abuses MessageProcessorMetaData
in MessagingYamlParsingDeployer by treating 'tasks:' as a queue, and Backgroundable
and TheTaskClassName as handler names. They get fleshed out in (Tasks|Backgroundable)Deployer
to be correct, but it still sucks.

Also, by changing MPMD to allow a concurrency of 0, I open another can of worms, since
it's know possible to specify MPMDs that can't be deployed, and it's stupid to have code
that checks to see if the MPMD is deployable.