Demand that HornetQ be configured before deploying MessageProcessors [TORQUE-262]
This fixes the issue where the MessageProcessor deploy fails with:

javax.jms.JMSSecurityException: Unable to validate user: null for check type CONSUME for address jms.queue./queues/torquebox/some/queue

It's caused by the HornetQCoreDeployment not yet being deployed,
which is what gives HornetQ its user/role data. They use an
incrementing integer as a suffix for the name (starting at 2),
so we have to depend on that name.