Now configuring caches according to a particular clustering mode [TORQUE-33]
We support the various infinispan modes: replicated, distributed, and
invalidation, which is the default.  So to have an asynchonous
replicated cache store:

  config.cache_store = :torque_box_store, {:mode => :replicated}

Which is equivalent to this:

  config.cache_store = ActiveSupport::Cache::TorqueBoxStore.new(:mode => :replicated)

The :sync option determines whether the write calls wait for acks from
other members of the cluster:

  config.cache_store = :torque_box_store, {:mode => :replicated, :sync => true}

It's also possible to configure multiple TB cache stores in your app,
each configured to be in a different clustering mode, but the
developer is responsible for uniquely naming them:

  ActiveSupport::Cache::TorqueBoxStore.new(:name => "models", :mode => :replicated)

Docs forthcoming!