We now support overriding any attribute of any deeply-nested section in a descriptor.
The only section we reference by name now is 'application', mostly cuz
we're flexible with its key names.  All other sections are merged in a
way that allows an external deployment descriptor to override any
attribute of a deeply-nested section.

We also introduced jruby in the test to make fixtures easier to create.