Support for durable topic subscribers [TORQUE-238]
Message processors can now be marked as durable in torquebox.yml
messaging: with durable: true. This only has effect on topic
processors, and is ignored if set on a queue processor (with a
warning in the log).

When a topic MP is marked as durable, HQ will create a durable
queue for that MP. The MP can then disconnect and later reconnect,
getting messages sent during its absence. For this to work, the
MP has to reconnect with the same client id and subscriber name.
We currently use the app name as the client id, and the MP name
(which contains the topic name and ruby clas name) as the
subscriber name.

There are no tests for this currently, since I don't see a way
to undeploy then redeploy an app inside arquillian.

It's also lacking docs atm, but those will be coming soon.