Ignore AST nodes we can't visit [TORQUE-328]
There are a few 1.9 AST nodes that JRuby doesn't accept visitors on,
so for now we skip them.

One of the two I've found so far is an optional argument to a method,
and the other is a lamdba node created using the new syntax:

    ->(arg1, arg2) { #body }

That means that inject() currently won't work inside a lamdba
defined that way, nor will it work as a default argument on a method:

    def foo(queue = inject('some_queue')) # won't be picked up by the analyzer

If either of these become an issue (or we discover more important nodes
we can't analyze), then we can look in to rewriting the analyzer to
use ParserSupport.

When we do hit a node we can't visit, we spit out a log.warn like so:

    JRuby doesn't support visiting node (OptArgNode:b 16, (LocalAsgnNode:b 16, (NilNode:nil 16))), skipping.