Ported override tests from junit to rspec.
They're still a tad awkward, though.  I'd prefer to use nested
describes instead of the "internal overrides" and "external overrides"
shared examples, but it seems that the :deploy call must reside in a
top-level describe in order for the global before/after config to
work.  I also would rather not have the redundant before(:each) calls
that set @home, but I couldn't get an additional shared example
consisting of only a :before call to work.  I think the shared example
stuff is still a work in progress.  I think what I need is
:include_examples, which will be available in rspec 2.6.