Rearrange torquebox-server gem contents a bit (TORQUE-434)
Our JBoss assembly is now at jboss/ instead of lib/jboss/. This gets
ride of the OOM error when installing torquebox-server gem with
default Xmx of 512m.

Also directly copy the assembly under target/torquebox-server/ instead
of to a top-level jboss/ directory so we don't have to copy it twice
and a 'mvn clean' cleans up after that as well.