Inject a service (defined via a services: block in your torquebox.yml) into other stuff.
    inject( "service:MyFancyService" )

To affect this change, The path of '.' has been removed from the default injection path
prefixes.  If it's present, and you have a config.ru in the root which injects
a service, the service itself picks up the injections from /* which results in a circular
dependency.

Thus, '.' has been removed from the default inject path prefixes, but explicitly added
back to the Rack web component injections.

Additionaly, for every service stood up, an ON_DEMAND .injectable service is also
installed.  Currently this returns the actual IRubyObject underneath the service.
It could, conceivably, return a RemoteRubyObjectProxy or something fancy to allow
removing or HA/failover behind a smart stub.