Add support for inject('runtime-injection-analyzer') which can then be used to inject_and_ananlyze(&block) to further load the block's own injections into the local Ruby interpreter's injection-registry, before actually executing that block.
To accomodate this, refactoring of our Injectables to work with a combination
of (ServiceTarget, DeploymentUnit) instead of the higher-level (DeploymentPhaseContext).
When using the DPC, we only grabbed the ST and DU from it anyhow.  At runtime
analysis, we do not have any access to a DPC.  But we surely have access to
a ServiceTarget and a DeploymentUnit.

Ta-da.