Adds a DSL for configuring TorqueBox [TORQUE-464]
This allows you to express anything you have in torquebox.yml via
ruby in torquebox.rb (except for any stomp stuff at this point). It
is undocumented currently - the best place to look for an example is
in the integs:

integration-tests/apps/rack/basic-torquebox-rb/config/torquebox.rb

This also lacks the planned mini-DSL for specifying cronspecs in
a similar fashion to the whenever gem.