We base64 encode all msg payloads now, since we can't tell whether Strings contain non-ascii data [TORQUE-473].
This affected the QueueBrowser since the text component of the message
is now encoded, so the QB now maps its real JMS messages to the TB
wrappers that know how to decode it.  And for good measure, we tossed
in a method_missing that delegates calls to the wrapped JMS message.