Initial cut of the STOMP/WebSockets support based upon Stilts.
A 'stomp:' section is supported in torquebox.yml, with a named section
for each stomplet, along with a route: and a class: parameter.

Routing rules follow normal Stomplet routing rules (see stilts.projectodd.org).
Class should specify a Ruby classname of a class which may be found in
APP/stomplets or APP/app/stomplets or right at the root of APP/.

Currently only 1 app containing stomplets may be deployed at a time,
due to lack of support for vhosting within the STOMP module.  This will
be remedied.

The Stomp server runs on port 8675, answering both pure STOMP and STOMP-over-websockets
on a unified server.