Bob determined that db rollbacks weren't occurring due to AR using a cached connection obtained before the tx was begun.
So for the time being, I introduced a TorqueBox.transaction{} that
will connection.reconnect! prior to yielding. We can add more logic
later to check for XA-ness and begin a transaction if need be. Still
coagulating.

Also, I put the XA session stuff back, enlisting the message receiving session in the transaction used for the db.