Recognize and convert a config/database.yml into a set of XA DataSources within the container, and secretly replace the config provided by ActiveRecord::Base.configurations with the JNDI name for grabbing the XA DataSource.
The JDBC driver is loaded from the user's own application gem, such
as jdbc-h2.gem or jdbc-postgres.gem.  The drivers and data-sources
that are deployed are available only within the scope of the application.
Other applications can, of course, use the datasource from JNDI, but
if the owning application is undeployed, behavior is undefined.

Additionally, predetermined-injections have been made available within
the initialization phase of a Ruby from a RubyRuntimeFactory, by injecting
them into the factory itself.