All InstanceFactory implementations must be thread-safe.
This should fix the deadlocks we've seen from ManagedPoolTest on CI
and Bob's machine - StringInstanceFactory wasn't thread-safe and could
possibly return identical instances which then get de-duped in the
Set<Instance> collections inside SimplePool.

I also found a potential issue with an unsychronized collection in
RuntimeContext so went ahead and wrapped that collection so it's
synchronized.