TORQUE-407 (redux): Guard the loading of the bulk of torquebox-messaging by the existance of javax.jms.* Java classes available to load.
If they fail, so would we, so might as well just warn-and-continue, given
this that affects outside-of-TorqueBox usage of our gems.

For instance, in a Rails app:

  ./script/console

This apparently will do the Bundler.setup() dance, and if a Gemfile
contains

  gem 'torquebox'

Then it will attempt to load a good many of our gems, including
torquebox-messaging.  In this context, JMS is not available.
Therefore, it failed.

This also addresses the issue reported by Vitorrio on torquebox-user@
mailing list around 25-Oct.

Also, ensure that 'thread' is required so that Mutex is actually
loadable within registry.rb.