Don't remove the XaDataSourceService if the datasource doesn't support XA (TORQUE-685)
Removing the XaDataSourceService triggers an entire chain of Services
to restart and ultimately leads to the null classloader seen in
TORQUE-685.

There was a flawed assumption that when an optional dependency is
removed that it doesn't trigger the dependent to stop. Not only does
it trigger the dependent to stop but the dependent also immediately
restarts because all its dependencies are satisfied even with the
optional dependency removed.