Remove VFSLoadService in favor of NonLeakingLoadService (TORQUE-775)
This is not only because I enjoy prefixing more classes with
NonLeaking, which I do, but also to greatly simplify our custom
LoadService logic. All VFS references are removed and no longer
needed, at least according to our unit and integration tests.

This is the first of several changes to get TorqueBox working with
JRuby 1.6.7 and 1.7.