Simplify newrelic background task monitoring.
This has the side effect of fixing TORQUE-780, since we were requiring the
newrelic_rpm gem before the pwd of the runtime was set by an initializer,
so it couldn't find the app's config. We now only enable background tracing
if the rpm gem has already been loaded, which will be the case if it's actually
being used in the app.