Remove dm-infinispan-adapter dependencies and pend all related integs.
Since we don't yet have dm-infinispan-adapter working with persistence
enabled, and we've removed all of the dm-infinispan-adapter
documentation, and nobody is even asking for it, let's also eliminate
the sizeable dependencies that enable searching the cache. Now, all of
the dm-infinispan-adapter integration specs are marked as pending, since
without their dependencies, they all fail. I wonder if we should
consider removing this feature altogether.