Unpend domain cache rollback test [TORQUE-855]
The 'repl' cache defined in domain.xml doesn't set a TMLookup
instance, so it isn't enlisted in the tx started in the test. Not
referring to an existing cache in torque_box_store.rb forces the
configuration of a new class and more importantly, sets a TMLookup. I
think it's good because it allows the rails developer to set the
desired replication mode in config/application.rb rather than
domain|standalone.xml.