Modify message processors to use middleware [TORQUE-899]
Message processors now provide a middleware method that returns
the default middleware. The current default middleware is just
WithTransaction, which replaces the old ProcessorWrapper.

This is just the initial implementation, and is subject to change
as I add more middleware to provide other functionality.

I also need to expose the middleware to torquebox.rb for adjustment.