Fix message procs and jobs statelessness in production mode (TORQUE-903)
We thought we were doing this before but turns out we weren't. This
change ensures that every message gets its own instance of the Ruby
MessageProcessor class and that every job gets its own instance of
the Ruby job class.

This makes production mode behave the same as development mode.