Don't bring up a web context until its runtime is ready (TORQUE-820)
Now a web context with eager pooling (the default) won't come up until
its Ruby runtime is started. With lazy pooling the context will still
come up before the runtime is started. This makes it much easier for
load balancers to know when an application is ready to serve requests
since as soon as the web context is available the application is ready.

Part of this change involed removing the start asynchronously option
from runtime pools. It was only used by the web context and not
exposed to the user so removing this should have little effect. All it
means is the web runtime pool is brought up before the application is
considered deployed instead of afterwards.