Support Rails page caching (TORQUE-824)
The StaticResourceServlet can now also serve files
created by Rails page caching. Note that right now it expects the
Rails page_cache_directory to be the same as your configured TorqueBox
static resource path and the default page cache extension to be
".html". If users request it, we can add the ability to customize
these defaults. As-is, Rails page caching should work out of the box
for most applications.