Fix Backgroundable usage with NewRelic (TORQUE-912)
Something.background.foo will no longer throw an error when NewRelic
is loaded in an application.

Fixing this led to discovery of another bug in applications with
NewRelic where it was possible for a method to get instrumented
multiple times leading to infinite loops and eventually stack level
too deep errors. This commit also fixes that issue.