Choose wiser between modifying or generating a Rails app (TORQUE-913)
Instead of going just off the root directory's existence, check for a
config/environment.rb to determine if a Rails application already
exists in the given directory or if we need to create one.

Also fix a bug in the way we specified gems for Rails 2.3.x versions
in our template.