Add a flag to manually start the HTTP connector for Heroku
When running on Heroku, we don't want to start the HTTP connector
until after an application is deployed. For now this is only used by
torquebox-lite which disables the HTTP connector in standalone.xml and
then sets org.torquebox.web.force_http_connector_start to force the
connector to start after the application gets deployed.