Minor cleanup of integ require statement
Only require 'active_support/core-ext/numeric/time' in the spec that
actually needs it. Otherwise, TorqueSpec's daemon tries to require
that file any time it gets loaded even if not running the spec that
needs that require. This leads to an error in sinatra_queues_spec or
any other non-Rails spec that uses remote_describe.