List all queues and topics (TORQUE-740)
This is a slight modification of the pull request for TORQUE-740 to
list all deployed queues and topics instead of trying to limit the
list to those deployed inside the current application. The logic for
that is surprisingly hard (how do we handle destinations created at
runtime instead of in deployment descriptors?) and added extra
complexity that I'd like to keep out for the first iteration of this
feature.

If users ask for a way to limit the list to only those deployed from a
specific application, we can revisit this in the future.