Updates to poorsmatic example application
All usage and mention of TorqueBox.transaction was removed - we don't
support XA with Datamapper so this wasn't working as intended anyway.

The database was also switched from Postgresql to Sqlite to simplify
setup locally. The TwitterService was modified to not error if the
user doesn't change the default dummy credentials - insteaad it just
logs a warning and doesn't start.