Pick up non-torquebox or torquebox-app-name security domains on boot.
I've added a lot of debug output to the security service to see what
is going on with picketlink and to be sure we're actually finding the
SecurityContext and AuthenticationManager from PicketLink. It seems that
we are, but alas, authentication consistently fails. The simple sinatra
app tests this with a ConfiguredIdentity in the "pork" security domain.

To test this with the integration tests, you'll need to change to the
integration-tests directory, run `mvn test-compile` to make sure you
have the latest and greatest, then add this stanza to standalone.xml in
the <security-domains> section (around line 399).

    <security-domain name='pork' cache-type='default'>
        <authentication>
            <login-module code='ConfiguredIdentity' flag='required'>
                <module-option name="userName" value="crunchy"/>
                <module-option name="principal" value="crunchy"/>
                <module-option name="password" value="bacon"/>
            </login-module>
        </authentication>
    </security-domain>