[TORQUE-877] Support for remote message processors
This commit adds back support for remote message processors. See messaging documentation for more info.

To connect to remote host a username and password authentication is required.
To lookup a destination it needs to be exported in the 'jboss/exported' JNDI naming context on the remote host to be visible outside of it.

DSL:

TorqueBox.configure do
  queue "/queue/remotequeue" do
    remote do
      host "somehost:4444"
      username "username"
      password "password"
    end

    processor SimpleProcessor do
      concurrency 2
    end
  end
end

YAML:

queues:
  /queue/remotequeue:
    remote:
      host: "somehost:4444"
      username: "username"
      password: "password"

messaging:
  /queue/remotequeue:
    SimpleProcessor:
      concurrency: 2

Additional changes:

1. Add 'org.projectodd' category to trace log for integs
2. Added documentation
3. Added integration test
4. Updated schema for YAML and DSL