Create a "global" service for destinations [TORQUE-1042]
When a topic or queue is deployed, it will now try to create a
top-level MSC service, along with a application-scoped service that
depends on the top-level one. If the top-level one already exists, the
application-scoped service alone is created.

The top-level service is ON_DEMAND, so as soon as it has no
dependents, it will stop(). Any new dependents will cause it to
start() again.

This makes it easier (and more reliable) for applications to share
destinations.

I may need to adjust the docs.