Depend on reference counting version of polyglot [TORQUE-1042]
Doing so requires the pointer service to be the only thing in an app
that depends on the global service. This means that destination
injections have to depend on the pointer service instead of the global
one, which means you can't inject a destination that hasn't been setup
for the application via torquebox.{yml,rb}. The one integ that tested
this behavior (sinatra_queues_rootless_spec) has been pended for now.

This isn't the final solution, but I want to see if ref counting
actually improves CI stability.