Perform all injection at runtime instead of deployment time (TORQUE-1047)
This is a fairly large changed to injection. Instead of a scanner and
analyzer that runs during the deployment, we now handle all injections
at runtime when the fetch method is called. This fixes a lot of
edge-case bugs with injection with regard to file locations, inactive
code, variables vs literal as arguments to fetch, etc.