Fix a race condition for multiple services of the same class (TORQUE-1058)
This fixes a race condition where it was possible that we were
reloading a service's .rb file (and thus redefining that service
class) while simultaneously trying to initialze an instance of that
class, leading to errors on service boot. This is a fairly rare issue
and only occurs with multiple services deployed in a single
application that use the same class.